requirements.txt
requests
Flask
python-dotenv
pyngrok
transformers
accelerate
torch
ffmpeg-python
revocalize
pydub
bark
scipy
